Module mpt_trie::debug_tools::diff

Diffing tools to compare two tries against each other. Useful when you want to find where the tries diverge from one each other.

There are a few considerations when implementing the logic to create a trie diff:

What should be reported when the trie node structures diverge (eg. two different node types proceeding a given common node)?

If there are multiple structural differences, how do we discover the smallest difference to report back?

If the node types between the tries (structure) are identical but some values are different, then these types of diffs are easy to detect and report the lowest difference. Structural differences are more challenging and a bit hard to report well. There are two approaches (only one currently is implemented) in how to detect structural differences:

Top-down search

Bottom-up search

These two searches are somewhat self-explanatory:

Top-down will find the highest point of a structural divergence and report it. If there are multiple divergences, then only the one that is the highest in the trie will be reported.

Bottom-up (not implemented) is a lot more complex to implement, but will attempt to find the smallest structural trie difference between the trie. If there are multiple differences, then this will likely be what you want to use.

Structs

DiffPoint

A point (node) between the two tries where the children differ.

NodeInfo

Meta information for a node in a trie.

TrieDiff

The difference between two Tries, represented as the highest point of a structural divergence.

Functions

create_diff_between_tries

Create a diff between two tries. Will perform both types of diff searches (top-down & bottom-up).