Module mpt_trie::debug_tools::query

Query tooling to report info on the path taken when searching down a trie with a given key.

Structs

DebugQuery

The payload to give to the query function. Construct this from the builder.

DebugQueryOutput

The result of a debug query contains information of the path used for searching for a key in the trie.

DebugQueryParams

Params controlling how much information is reported in the query output.

DebugQueryParamsBuilder

A wrapper for DebugQueryParams.

Functions

get_path_from_query

Get debug information on the path taken when querying a key in a given trie.