Module mpt_trie::nibbles

Define Nibbles and how to convert bytes, hex prefix encodings and strings into nibbles.

Structs

Nibbles

A sequence of nibbles which is used as the key type into PartialTrie.

NibblesIntern

Little-endian large integer type Used for the internal representation of a sequence of nibbles. The choice of [u64; 5] accommodates the 260-bit key requirement efficiently by leveraging 64-bit instructions for performance, while providing an additional u64 to handle the overflow case beyond the 256-bit capacity of [u64; 4].

StrToNibblesError

An error encountered when converting a string to a sequence of nibbles.

Enums

BytesToNibblesError

Errors encountered when converting from Bytes to Nibbles.

FromHexPrefixError

Errors encountered when converting to hex prefix encoding to nibbles.

NibblesToTypeError

Error type for conversion.

Traits

ToNibbles

Because there are two different ways to convert to Nibbles, we don’t want to rely on From. Instead, we’ll define a new trait that defines both conversions.

Type Aliases

Nibble

A Nibble has 4 bits and is stored as u8.