Struct mpt_trie::nibbles::Nibbles

pub struct Nibbles {
    pub count: usize,
    pub packed: NibblesIntern,
}

A sequence of nibbles which is used as the key type into PartialTrie.

Generally, if you’re constructing keys from actual trie data, you probably will be working with U256s and H256s both of which Nibbles has a From implementation for.

It’s important to note that leading 0 bits are part of a key. For example:

let n1 = Nibbles::from_str("0x123").unwrap();
let n2 = Nibbles::from_str("0x0123").unwrap();

assert_ne!(n1, n2); // These are different keys

Also note by default, converting to Nibbles does not pad to the nearest byte like other trie libraries generally do. If you need this behavior, you can construct Nibbles like this:

let padded = 0x123_u64.to_nibbles_byte_padded();
assert_eq!(format!("{:x}", padded), "0x0123");

Note that for the time being, Nibbles is limited to key lengths no longer than 256 bits. While we could support arbitrarily long keys, tries in Ethereum never have keys longer than 256 bits. Because of this, we decided to create a minor optimization by restricting max key sizes to 256 bits.

Finally, note that due to the limitations initializing from an integer, when creating a key directly from an integer, there is no way to know if a leading 0 was passed in.

let n1 = Nibbles::from(0x123_u64);
let n2 = Nibbles::from(0x00000000123_u64); // Use `from_str` or construct `Nibbles` explicitly instead here.

assert_eq!(n1, n2);
assert!(Nibbles::from(0x00000000_u64).is_empty());

Fields
count: usize

The number of nibbles in this sequence.

packed: NibblesIntern

A packed encoding of these nibbles. Only the first (least significant) 4 * count bits are used. The rest are unused and should be zero.

Implementations

impl Nibbles

pub fn new() -> Self

Create Nibbles that is empty.

Note that mean that the key size is 0 and does not mean that the key contains the 0 Nibble.

pub fn from_bytes_be(bytes: &[u8]) -> Result<Self, BytesToNibblesError>

Creates Nibbles from big endian bytes.

Returns an error if the byte slice is empty or is longer than 32 bytes.

pub fn from_bytes_le(bytes: &[u8]) -> Result<Self, BytesToNibblesError>

Creates Nibbles from little endian bytes.

Returns an error if the byte slice is empty or is longer than 32 bytes.

pub fn from_nibble(n: Nibble) -> Self

Creates a new Nibbles from a single Nibble.

Panics

Panics if the nibble is > 0xf.

pub fn get_nibble(&self, idx: usize) -> Nibble

Gets the nth proceeding nibble. The front Nibble is at idx 0.

Panics

Panics if idx is out of range.

pub fn pop_next_nibble_front(&mut self) -> Nibble

Pops the nibble at the front (the next nibble).

Panics

Panics if the Nibbles is empty.

pub fn pop_next_nibble_back(&mut self) -> Nibble

Pops the nibble at the back (the last nibble).

Panics

Panics if the Nibbles is empty.

pub fn get_next_nibbles(&self, n: usize) -> Nibbles

Gets the next n nibbles.

Panics

Panics if n is larger than the number of nibbles contained.

pub fn pop_nibbles_front(&mut self, n: usize) -> Nibbles

Pops the next n nibbles from the front.

Panics

Panics if n is larger than the number of nibbles contained.

pub fn pop_nibbles_back(&mut self, n: usize) -> Nibbles

Pops the next n nibbles from the back.

Panics

Panics if n is larger than the number of nibbles contained.

pub fn push_nibble_front(&mut self, n: Nibble)

Pushes a nibble to the front.

Panics

Panics if appending the Nibble causes an overflow (total nibbles > 64).

pub fn push_nibble_back(&mut self, n: Nibble)

Pushes a nibble to the back.

Panics

Panics if appending the Nibble causes an overflow (total nibbles > 64).

pub fn push_nibbles_front(&mut self, n: &Self)

Appends Nibbles to the front.

Panics

Panics if appending the Nibbles causes an overflow (total nibbles > 64).

pub fn push_nibbles_back(&mut self, n: &Self)

Appends Nibbles to the back.

Panics

Panics if appending the Nibbles causes an overflow (total nibbles > 64).

pub fn get_nibble_range(&self, range: Range<usize>) -> Nibbles

Gets the nibbles at the range specified, where 0 is the next nibble.

Panics

Panics if range.end is outside of the current Nibbles.

pub fn truncate_n_nibbles_front(&self, n: usize) -> Nibbles

Drops the next n proceeding nibbles without mutation.

If we truncate more nibbles that there are, we will just return the empty nibble.

pub fn truncate_n_nibbles_back(&self, n: usize) -> Nibbles

Drops the last n nibbles without mutation.

If we truncate more nibbles that there are, we will just return the empty nibble.

pub fn truncate_n_nibbles_front_mut(&mut self, n: usize)

Drop the next n proceeding nibbles.

If we truncate more nibbles that there are, we will just return the empty nibble.

pub fn truncate_n_nibbles_back_mut(&mut self, n: usize)

Drop the last n nibbles.

If we truncate more nibbles that there are, we will just return the empty nibble.

pub const fn is_empty(&self) -> bool

Returns whether or not this Nibbles contains actual nibbles. (If count is set to 0)

pub fn nibbles_are_identical_up_to_smallest_count( &self, other: &Nibbles, ) -> bool

Checks if two given Nibbles are identical up to the shorter of the two Nibbles.

pub fn split_at_idx(&self, idx: usize) -> (Nibbles, Nibbles)

Splits the Nibbles at the given index, returning two Nibbles. Specifically, if 0x1234 is split at 1, we get 0x1 and 0x234.

Panics

Panics if the idx is out of range.

pub fn split_at_idx_prefix(&self, idx: usize) -> Nibbles

Split the Nibbles at the given index but only return the prefix.

Panics

Panics if the idx is out of range.

pub fn split_at_idx_postfix(&self, idx: usize) -> Nibbles

Split the Nibbles at the given index but only return the postfix.

Panics

Panics if the idx is out of range.

pub fn merge_nibble(&self, post: Nibble) -> Nibbles

Merge a single Nibble with a Nibbles. self will be the prefix.

Panics

Panics if merging the Nibble causes an overflow (total nibbles > 64).

pub fn merge_nibbles(&self, post: &Nibbles) -> Nibbles

Merge two Nibbles together. self will be the prefix.

Panics

Panics if merging the Nibbles causes an overflow (total nibbles > 64).

pub fn reverse(&self) -> Nibbles

Reverses the Nibbles such that the last Nibble is now the first Nibble.

pub fn find_nibble_idx_that_differs_between_nibbles_different_lengths( n1: &Nibbles, n2: &Nibbles, ) -> usize

Finds the nibble idx that differs between two nibbles. If there is no difference, returns 1 + the last index.

pub fn find_nibble_idx_that_differs_between_nibbles_equal_lengths( n1: &Nibbles, n2: &Nibbles, ) -> usize

Finds the nibble index that differs between two Nibbles of equal length. If there is no difference, returns 1 + the last index.

Panics

Panics if both Nibbles are not the same length.

pub fn to_hex_prefix_encoding(&self, is_leaf: bool) -> Bytes

Converts Nibbles to hex-prefix encoding (AKA “compact”). This appends an extra nibble to the end which encodes if the node is even and if it’s a leaf (terminator) or not.

pub fn from_hex_prefix_encoding( hex_prefix_bytes: &[u8], ) -> Result<Self, FromHexPrefixError>

Converts a hex prefix byte string (“AKA “compact”) into Nibbles.

pub const fn min_bytes(&self) -> usize

Returns the minimum number of bytes needed to represent these Nibbles.

pub fn get_num_nibbles_in_key(k: &NibblesIntern) -> usize

Returns the minimum number of nibbles needed to represent a U256 key.

pub fn bytes_be(&self) -> Vec<u8>

Returns the nibbles bytes in big-endian format.

pub fn from_h256_be(v: H256) -> Self

Creates Nibbles from a big endian H256.

pub fn from_h256_le(v: H256) -> Self

Creates Nibbles from a little endian H256.

pub const fn as_byte_slice(&self) -> &[u8]

Returns a slice of the internal bytes of packed nibbles. Only the relevant bytes (up to count nibbles) are considered valid.