Trait mpt_trie::nibbles::ToNibbles

pub trait ToNibbles {
    // Required method
    fn to_nibbles(self) -> Nibbles;

    // Provided method
    fn to_nibbles_byte_padded(self) -> Nibbles
       where Self: Sized { ... }
}

Because there are two different ways to convert to Nibbles, we don’t want to rely on From. Instead, we’ll define a new trait that defines both conversions.

Required Methods

fn to_nibbles(self) -> Nibbles

Convert the type to a sequence of nibbles.

Note that this will create Nibbles with a Nibble count that is accurate down to the nibble. For example, passing in 0x123 has 3 Nibbles and is not padded to the nearest byte (in which case it would have 4 Nibbles).

Provided Methods

fn to_nibbles_byte_padded(self) -> Nibbles
where Self: Sized,

Convert the type to a sequence of nibbles but pad to the nearest byte.