Trait mpt_trie::partial_trie::PartialTrie

pub trait PartialTrie:
    Clone
    + Debug
    + Default
    + DerefMut<Target = Node<Self>>
    + Eq
    + TrieNodeIntern {
    // Required methods
    fn new(n: Node<Self>) -> Self;
    fn new_with_strategy(n: Node<Self>, strategy: OnOrphanedHashNode) -> Self;
    fn insert<K, V>(&mut self, k: K, v: V) -> TrieOpResult<()>
       where K: Into<Nibbles>,
             V: Into<ValOrHash>;
    fn extend<K, V, I>(&mut self, nodes: I) -> TrieOpResult<()>
       where K: Into<Nibbles>,
             V: Into<ValOrHash>,
             I: IntoIterator<Item = (K, V)>;
    fn get<K>(&self, k: K) -> Option<&[u8]>
       where K: Into<Nibbles>;
    fn delete<K>(&mut self, k: K) -> TrieOpResult<Option<Vec<u8>>>
       where K: Into<Nibbles>;
    fn hash(&self) -> H256;
    fn items(&self) -> impl Iterator<Item = (Nibbles, ValOrHash)>;
    fn keys(&self) -> impl Iterator<Item = Nibbles>;
    fn values(&self) -> impl Iterator<Item = ValOrHash>;
    fn contains<K>(&self, k: K) -> bool
       where K: Into<Nibbles>;
}

A trait for any types that are Tries.

Required Methods

fn new(n: Node<Self>) -> Self

Creates a new partial trie from a node.

fn new_with_strategy(n: Node<Self>, strategy: OnOrphanedHashNode) -> Self

Creates a new partial trie from a node with a provided collapse strategy.

fn insert<K, V>(&mut self, k: K, v: V) -> TrieOpResult<()>
where K: Into<Nibbles>, V: Into<ValOrHash>,

Inserts a node into the trie.

fn extend<K, V, I>(&mut self, nodes: I) -> TrieOpResult<()>
where K: Into<Nibbles>, V: Into<ValOrHash>, I: IntoIterator<Item = (K, V)>,

Add more nodes to the trie through an iterator

fn get<K>(&self, k: K) -> Option<&[u8]>
where K: Into<Nibbles>,

Get a node if it exists in the trie.

fn delete<K>(&mut self, k: K) -> TrieOpResult<Option<Vec<u8>>>
where K: Into<Nibbles>,

Deletes a Leaf node or Branch value field if it exists.

To agree with Ethereum specs, deleting nodes does not result in the trie removing nodes that are redundant after deletion. For example, a Branch node that is completely empty after all of its children are deleted is not pruned. Also note:

Deleted leaves are replaced with Empty nodes.

Deleted branch values are replaced with empty Vecs.

Panics

If a Hash node is traversed, a panic will occur. Since Hash nodes are meant for parts of the trie that are not relevant, traversing one means that a Hash node was created that potentially should not have been.

fn hash(&self) -> H256

Get the hash for the node.

fn items(&self) -> impl Iterator<Item = (Nibbles, ValOrHash)>

Returns an iterator over the trie that returns all key/value pairs for every Leaf and Hash node.

fn keys(&self) -> impl Iterator<Item = Nibbles>

Returns an iterator over the trie that returns all keys for every Leaf and Hash node.

fn values(&self) -> impl Iterator<Item = ValOrHash>

Returns an iterator over the trie that returns all values for every Leaf and Hash node.

fn contains<K>(&self, k: K) -> bool
where K: Into<Nibbles>,

Returns true if the trie contains an element with the given key.

Object Safety

This trait is not object safe.