Module mpt_trie::trie_ops

Defines various operations for PartialTrie.

Structs

PartialTrieIter

An iterator that ranges over all the leafs and hash nodes of the trie, in lexicographic order.

Enums

TrieOpError

An error type for trie operation.

ValOrHash

An “entry” in a PartialTrie.

Type Aliases

TrieOpResult

Stores the result of trie operations. Returns a TrieOpError upon failure.