Module mpt_trie::utils

Various types and logic that don’t fit well into any other module.

Structs

TriePath

A vector of path segments representing a path in the trie.

Enums

TrieNodeType

Simplified trie node type to make logging cleaner.

TrieSegment

Minimal key information of “segments” (nodes) used to construct trie “traces” of a trie query. Unlike TrieNodeType, this type also contains the key piece of the node if applicable (eg. Node::Empty & Node::Hash do not have associated key pieces).

Traits

IntoTrieKey

Trait for a type that can be converted into a trie key (Nibbles).

TryFromIterator

Conversion from an Iterator within an allocator.