Struct mpt_trie::partial_trie::HashedPartialTrie

pub struct HashedPartialTrie { /* private fields */ }

A partial trie that lazily caches hashes for each node as needed. If you are doing frequent hashing of node, you probably want to use this Trie variant.